[FEAT]토큰정보 불일치 해결

+++ src/main/java/com/example/ot/config/security/jwt/JwtProvider.java
@@ -20,7 +20,7 @@ public class JwtProvider {
     }
 
     // 토큰 생성
-    public String generateAccessToken(Map<String, Object> claims, int seconds) {
+    public String generateAccessToken(Map<String, Object> claims, long seconds) {
         long now = new Date().getTime();
         Date accessTokenExpiresIn = new Date(now + 1000L * seconds);
